security: move computeChangeMap out of "use server" (IDOR)
computeChangeMap was an unguarded server action: any authenticated user could
invoke it RPC-style with any buildId, reading and recomputing another team's
build data on their dime. It now lives in @/lib/change-map/compute as a plain
function; the remaining server action and the API routes call it behind their
own ownership checks.
